get eth price and sell token prices
